# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Client Escalation Calls API
  description: |
    Agentic AI Payment Collection System API
    
    This API provides endpoints for managing customers, payment records, 
    collection campaigns, tasks, and AI agents for automated debt collection.
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain a token via the `/auth/login` endpoint.
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15-minute window per IP.
    
    ## Error Handling
    All errors follow a consistent format with `error` and optional `details` fields.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Get paginated list of customers with optional filtering
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search by company name, contact name, or email
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by risk level
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [companyName, createdAt, updatedAt]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CustomerSummary'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [Customers]
      summary: Create customer
      description: Create a new customer record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Customer with email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      description: Get paginated list of payment records
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, overdue, paid, partial, cancelled]
        - name: overdueOnly
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [dueDate, amount, createdAt]
            default: dueDate
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Payment records list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Payment'

    post:
      tags: [Payments]
      summary: Create payment record
      description: Create a new payment record (invoice)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment record created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      description: Get paginated list of collection campaigns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, paused, completed, cancelled]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt, startDate]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CampaignSummary'

    post:
      tags: [Campaigns]
      summary: Create campaign
      description: Create a new collection campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      description: Get paginated list of agent tasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_progress, completed, failed, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [send_email, make_call, send_sms, research_customer, escalate, follow_up]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: assignedAgentId
          in: query
          schema:
            type: string
            format: uuid
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, dueAt, priority]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskSummary'

    post:
      tags: [Tasks]
      summary: Create task
      description: Create a new agent task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      description: Get paginated list of AI agents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [email, phone, sms, research]
        - name: status
          in: query
          schema:
            type: string
            enum: [idle, active, busy, error, offline]
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Agent'

    post:
      tags: [Agents]
      summary: Create agent
      description: Create a new AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    UnauthorizedError:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: string
          description: Specific validation failure message
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            totalCount:
              type: integer
            totalPages:
              type: integer
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string

    # ============================================================================
    # Health & Stats
    # ============================================================================
...
    CreateCustomerRequest:
      type: object
      required: [companyName, contactName, email]
      properties:
        companyName:
          type: string
          maxLength: 255
        contactName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        mobile:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        address:
          $ref: '#/components/schemas/Address'
        preferredContactMethod:
          type: string
          enum: [email, phone, sms]
          default: email
        tags:
          type: array
          items:
            type: string
...
    CustomerSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        responseRate:
          type: number
        createdAt:
          type: string
          format: date-time
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        mobile:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        preferredContactMethod:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreatePaymentRequest:
      type: object
      required: [customerId, amount, dueDate, invoiceNumber]
      properties:
        customerId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          default: USD
          minLength: 3
          maxLength: 3
        dueDate:
          type: string
          format: date
        invoiceNumber:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        amount:
          type: number
        currency:
          type: string
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, overdue, paid, partial, cancelled]
        invoiceNumber:
          type: string
        description:
          type: string
        daysOverdue:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EscalationStep:
      type: object
      required: [stepNumber, channel, template]
      properties:
        stepNumber:
          type: integer
          minimum: 1
        channel:
          type: string
          enum: [email, phone, sms]
        template:
          type: string
        delayHours:
          type: integer
          minimum: 0
          default: 0
        maxAttempts:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
    CampaignConfig:
      type: object
      properties:
        maxDailyContacts:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
        cooldownPeriod:
          type: integer
          minimum: 1
          default: 4
        respectDoNotContact:
          type: boolean
          default: true
        timezone:
          type: string
          default: "America/New_York"
    CreateCampaignRequest:
      type: object
      required: [name, customerId, escalationSteps]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        customerId:
          type: string
          format: uuid
        paymentRecordId:
          type: string
          format: uuid
        escalationSteps:
          type: array
          items:
            $ref: '#/components/schemas/EscalationStep'
          minItems: 1
        config:
          $ref: '#/components/schemas/CampaignConfig'
        startDate:
          type: string
          format: date-time
...
    CampaignSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, cancelled]
        currentStep:
          type: integer
        startDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Campaign:
      allOf:
        - $ref: '#/components/schemas/CampaignSummary'
        - type: object
          properties:
            description:
              type: string
            escalationSteps:
              type: array
              items:
                $ref: '#/components/schemas/EscalationStep'
            config:
              $ref: '#/components/schemas/CampaignConfig'
            results:
              type: object
            endDate:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    TaskContext:
      type: object
      properties:
        paymentRecordId:
          type: string
          format: uuid
        messageTemplate:
          type: string
        metadata:
          type: object
    CreateTaskRequest:
      type: object
      required: [type, customerId, campaignId]
      properties:
        type:
          type: string
          enum: [send_email, make_call, send_sms, research_customer, escalate, follow_up]
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        customerId:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        dueAt:
          type: string
          format: date-time
        maxAttempts:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
        context:
          $ref: '#/components/schemas/TaskContext'
...
    TaskSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        priority:
          type: string
        status:
          type: string
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        campaignId:
          type: string
          format: uuid
        campaignName:
          type: string
        assignedAgentId:
          type: string
          format: uuid
        agentType:
          type: string
        attempts:
          type: integer
        maxAttempts:
          type: integer
        dueAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Task:
      allOf:
        - $ref: '#/components/schemas/TaskSummary'
        - type: object
          properties:
            context:
              type: object
            completedAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    AgentConfig:
      type: object
      properties:
        maxConcurrentTasks:
          type: integer
          minimum: 1
          maximum: 20
          default: 5
        workingHours:
          type: object
          properties:
            start:
              type: string
              pattern: '^\d{2}:\d{2}$'
              default: "09:00"
            end:
              type: string
              pattern: '^\d{2}:\d{2}$'
              default: "17:00"
        timezone:
          type: string
          default: "America/New_York"
        skills:
          type: array
          items:
            type: string
    CreateAgentRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [email, phone, sms, research]
        capabilities:
          type: array
          items:
            type: string
        config:
          $ref: '#/components/schemas/AgentConfig'
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [email, phone, sms, research]
        status:
          type: string
          enum: [idle, active, busy, error, offline]
        capabilities:
          type: array
          items:
            type: string
        currentTasks:
          type: integer
        performance:
          type: object
        config:
          $ref: '#/components/schemas/AgentConfig'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...